Split codec.js into parts
