+++ js/noms/src/database-test.js
@@ -15,7 +15,7 @@ import Database from './database.js';
 import {notNull} from './assert.js';
 import List from './list.js';
 import Struct, {newStruct} from './struct.js';
-import {encodeValue} from './codec.js';
+import encodeValue from './encode-value.js';
 import NomsSet from './set.js'; // namespace collision with JS Set
 import {invariant} from './assert.js';
 import {Kind} from './noms-kind.js';
