+++ js/noms/src/remote-batch-store-test.js
@@ -7,7 +7,7 @@
 import {suite, test} from 'mocha';
 import {assert} from 'chai';
 import makeRemoteBatchStoreFake from './remote-batch-store-fake.js';
-import {encodeValue} from './codec.js';
+import encodeValue from './encode-value.js';
 
 suite('BatchStore', () => {
   test('get after schedulePut works immediately', async () => {
