+++ js/noms/src/value-store-test.js
@@ -10,7 +10,7 @@ import MemoryStore from './memory-store.js';
 import {BatchStoreAdaptor} from './batch-store.js';
 import ValueStore from './value-store.js';
 import List from './list.js';
-import {encodeValue} from './codec.js';
+import encodeValue from './encode-value.js';
 import {equals} from './compare.js';
 import Hash from './hash.js';
 import {getHash} from './get-hash.js';
